% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scale_suffix.R
\name{add_scale_suffix}
\alias{add_scale_suffix}
\title{Add appropriate y-axis scale suffix based on magnitude}
\usage{
add_scale_suffix(p, max_abs_value)
}
\arguments{
\item{p}{ggplot object}

\item{max_abs_value}{Maximum absolute value to determine scale}
}
\value{
Modified ggplot object
}
\description{
Add appropriate y-axis scale suffix based on magnitude
}
\keyword{internal}
