% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_roic_data.R
\name{build_daily_roic}
\alias{build_daily_roic}
\title{Build Daily ROIC for Single Ticker}
\usage{
build_daily_roic(price_data, ttm_data, ticker)
}
\arguments{
\item{price_data}{Daily price data}

\item{ttm_data}{Quarterly TTM data with per-share columns}

\item{ticker}{Ticker symbol}
}
\value{
Data frame with daily ROIC
}
\description{
Build Daily ROIC for Single Ticker
}
\keyword{internal}
