% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_buyback_yield.R
\name{calculate_buyback_yield}
\alias{calculate_buyback_yield}
\title{Calculate Buyback Yield}
\usage{
calculate_buyback_yield(buyback_per_share, price)
}
\arguments{
\item{buyback_per_share}{TTM buyback per share (numeric vector)}

\item{price}{Stock price (numeric vector)}
}
\value{
Buyback yield as decimal (numeric vector)
}
\description{
Calculate Buyback Yield
}
\keyword{internal}
