% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dividend_yield.R
\name{calculate_dividend_yield}
\alias{calculate_dividend_yield}
\title{Calculate Dividend Yield}
\usage{
calculate_dividend_yield(dividend_per_share, price)
}
\arguments{
\item{dividend_per_share}{TTM dividend per share (numeric vector)}

\item{price}{Stock price (numeric vector)}
}
\value{
Dividend yield as decimal (numeric vector)
}
\description{
Calculate Dividend Yield
}
\keyword{internal}
