% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ev_to_ebitda.R
\name{calculate_ev_to_ebitda}
\alias{calculate_ev_to_ebitda}
\title{Calculate EV-to-EBITDA Ratio}
\usage{
calculate_ev_to_ebitda(ev_per_share, ebitda_per_share)
}
\arguments{
\item{ev_per_share}{Enterprise value per share (numeric vector)}

\item{ebitda_per_share}{EBITDA per share (numeric vector)}
}
\value{
EV/EBITDA ratio (numeric vector)
}
\description{
Calculate EV-to-EBITDA Ratio
}
\keyword{internal}
