% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ev_to_nopat.R
\name{calculate_ev_to_nopat}
\alias{calculate_ev_to_nopat}
\title{Calculate EV-to-NOPAT Ratio}
\usage{
calculate_ev_to_nopat(ev_per_share, nopat_per_share)
}
\arguments{
\item{ev_per_share}{Enterprise value per share (numeric vector)}

\item{nopat_per_share}{NOPAT per share (numeric vector)}
}
\value{
EV/NOPAT ratio (numeric vector)
}
\description{
Calculate EV-to-NOPAT Ratio
}
\keyword{internal}
