% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fcf.R
\name{calculate_fcf}
\alias{calculate_fcf}
\title{Calculate Free Cash Flow}
\usage{
calculate_fcf(operating_cashflow, capital_expenditures)
}
\arguments{
\item{operating_cashflow}{Operating cash flow vector}

\item{capital_expenditures}{Capital expenditures vector (positive values)}
}
\value{
Numeric vector of FCF values
}
\description{
Vectorized function for use inside dplyr::mutate().
}
