% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nopat.R
\name{calculate_nopat}
\alias{calculate_nopat}
\title{Calculate NOPAT (Net Operating Profit After Tax)}
\usage{
calculate_nopat(ebit, dep_amort, depreciation, tax_rate = 0.2375)
}
\arguments{
\item{ebit}{EBIT vector}

\item{dep_amort}{Depreciation and amortization vector}

\item{depreciation}{Depreciation vector (to isolate amortization)}

\item{tax_rate}{Tax rate (default: 0.2375)}
}
\value{
Numeric vector of NOPAT values
}
\description{
Vectorized function for use inside dplyr::mutate().
}
