% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_price_to_ebit.R
\name{calculate_price_to_ebit}
\alias{calculate_price_to_ebit}
\title{Calculate Price-to-EBIT Ratio}
\usage{
calculate_price_to_ebit(price, ebit_per_share)
}
\arguments{
\item{price}{Stock price (numeric vector)}

\item{ebit_per_share}{EBIT per share (numeric vector)}
}
\value{
P/EBIT ratio (numeric vector)
}
\description{
Calculate Price-to-EBIT Ratio
}
\keyword{internal}
