% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_price_to_gross_profit.R
\name{calculate_price_to_gross_profit}
\alias{calculate_price_to_gross_profit}
\title{Calculate Price-to-Gross-Profit Ratio}
\usage{
calculate_price_to_gross_profit(price, gross_profit_per_share)
}
\arguments{
\item{price}{Stock price (numeric vector)}

\item{gross_profit_per_share}{Gross profit per share (numeric vector)}
}
\value{
P/Gross Profit ratio (numeric vector)
}
\description{
Calculate Price-to-Gross-Profit Ratio
}
\keyword{internal}
