% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_price_to_sales.R
\name{calculate_price_to_sales}
\alias{calculate_price_to_sales}
\title{Calculate Price-to-Sales Ratio}
\usage{
calculate_price_to_sales(price, revenue_per_share)
}
\arguments{
\item{price}{Stock price (numeric vector)}

\item{revenue_per_share}{Revenue per share (numeric vector)}
}
\value{
P/S ratio (numeric vector)
}
\description{
Calculate Price-to-Sales Ratio
}
\keyword{internal}
