% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_roa.R
\name{calculate_roa}
\alias{calculate_roa}
\title{Calculate Return on Assets}
\usage{
calculate_roa(income, assets)
}
\arguments{
\item{income}{Numeric vector of income values (e.g., net income, NOPAT)}

\item{assets}{Numeric vector of total asset values}
}
\value{
Numeric vector of ROA as decimals
}
\description{
Vectorized calculation of ROA as a decimal.
}
