% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_valuation_data.R
\name{calculate_sector_ev_nopat}
\alias{calculate_sector_ev_nopat}
\title{Calculate Sector Average EV/NOPAT}
\usage{
calculate_sector_ev_nopat(price_data, ttm_data, sector_tickers)
}
\arguments{
\item{price_data}{Daily price data}

\item{ttm_data}{Quarterly TTM data}

\item{sector_tickers}{Tickers in the sector}
}
\value{
Data frame with daily sector EV/NOPAT ratio
}
\description{
Calculate Sector Average EV/NOPAT
}
\keyword{internal}
