% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_roic_data.R
\name{calculate_sector_roic}
\alias{calculate_sector_roic}
\title{Calculate Sector Median ROIC}
\usage{
calculate_sector_roic(price_data, ttm_data, sector_tickers)
}
\arguments{
\item{price_data}{Daily price data}

\item{ttm_data}{Quarterly TTM data}

\item{sector_tickers}{Tickers in the sector}
}
\value{
Data frame with daily sector median ROIC
}
\description{
Calculate Sector Median ROIC
}
\keyword{internal}
