% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shareholder_yield.R
\name{calculate_shareholder_yield}
\alias{calculate_shareholder_yield}
\title{Calculate Shareholder Yield}
\usage{
calculate_shareholder_yield(dividend_per_share, buyback_per_share, price)
}
\arguments{
\item{dividend_per_share}{TTM dividend per share (numeric vector)}

\item{buyback_per_share}{TTM buyback per share (numeric vector)}

\item{price}{Stock price (numeric vector)}
}
\value{
Shareholder yield as decimal (numeric vector)
}
\description{
Sum of dividend yield and buyback yield.
}
\keyword{internal}
