% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital_turnover.R
\name{capital_turnover}
\alias{capital_turnover}
\title{Calculate Capital Turnover}
\usage{
capital_turnover(revenue, invested_capital)
}
\arguments{
\item{revenue}{Revenue (numeric vector)}

\item{invested_capital}{Invested capital (numeric vector)}
}
\value{
Capital turnover ratio (numeric vector)
}
\description{
Calculate Capital Turnover
}
\keyword{internal}
