% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_old_logs.R
\name{clean_old_logs}
\alias{clean_old_logs}
\title{Clean Old Log Files}
\usage{
clean_old_logs(log_dir, retention_days)
}
\arguments{
\item{log_dir}{Directory containing log files}

\item{retention_days}{Number of days to retain}
}
\value{
Invisible NULL
}
\description{
Removes log files older than retention period.
}
\keyword{internal}
