% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_debt.R
\name{cost_of_debt}
\alias{cost_of_debt}
\title{Calculate Cost of Debt}
\usage{
cost_of_debt(interest_expense, avg_debt)
}
\arguments{
\item{interest_expense}{Interest expense (numeric vector)}

\item{avg_debt}{Average debt (numeric vector)}
}
\value{
Cost of debt as decimal (numeric vector)
}
\description{
Calculate Cost of Debt
}
\keyword{internal}
