% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_to_ebitda.R
\name{debt_to_ebitda}
\alias{debt_to_ebitda}
\title{Calculate Debt to EBITDA Ratio}
\usage{
debt_to_ebitda(total_debt, ebitda)
}
\arguments{
\item{total_debt}{Total debt (numeric vector)}

\item{ebitda}{EBITDA (numeric vector)}
}
\value{
Debt to EBITDA ratio (numeric vector)
}
\description{
Calculate Debt to EBITDA Ratio
}
\keyword{internal}
