% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_interaction.R
\name{distribute_interaction}
\alias{distribute_interaction}
\title{Distribute Interaction Term Proportionally}
\usage{
distribute_interaction(contrib_a, contrib_b, interaction)
}
\arguments{
\item{contrib_a}{Numeric vector of first contribution values}

\item{contrib_b}{Numeric vector of second contribution values}

\item{interaction}{Numeric vector of interaction term values}
}
\value{
List with a_adjusted and b_adjusted vectors
}
\description{
Distributes an interaction term between two contributions proportionally to their magnitudes.
}
