% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcf_conversion.R
\name{fcf_conversion}
\alias{fcf_conversion}
\title{Calculate FCF Conversion}
\usage{
fcf_conversion(fcf, nopat)
}
\arguments{
\item{fcf}{Free cash flow (numeric vector)}

\item{nopat}{Net operating profit after tax (numeric vector)}
}
\value{
FCF conversion ratio (numeric vector)
}
\description{
Calculate FCF Conversion
}
\keyword{internal}
