% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_logger.R
\name{get_app_logger}
\alias{get_app_logger}
\title{Get Application Logger}
\usage{
get_app_logger()
}
\value{
lgr Logger object
}
\description{
Returns the configured application logger.
}
\keyword{internal}
