% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groic.R
\name{groic}
\alias{groic}
\title{Calculate Gross Return on Invested Capital (GROIC)}
\usage{
groic(gross_profit, invested_capital)
}
\arguments{
\item{gross_profit}{Gross profit (numeric vector)}

\item{invested_capital}{Invested capital (numeric vector)}
}
\value{
GROIC as decimal (numeric vector)
}
\description{
Calculate Gross Return on Invested Capital (GROIC)
}
\keyword{internal}
