% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gross_margin.R
\name{gross_margin}
\alias{gross_margin}
\title{Calculate Gross Margin}
\usage{
gross_margin(gross_profit, revenue)
}
\arguments{
\item{gross_profit}{Gross profit (numeric vector)}

\item{revenue}{Revenue (numeric vector)}
}
\value{
Gross margin as decimal (numeric vector)
}
\description{
Calculate Gross Margin
}
\keyword{internal}
