% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_coverage.R
\name{interest_coverage}
\alias{interest_coverage}
\title{Calculate Interest Coverage Ratio}
\usage{
interest_coverage(ebit, interest_expense)
}
\arguments{
\item{ebit}{Earnings before interest and taxes (numeric vector)}

\item{interest_expense}{Interest expense (numeric vector)}
}
\value{
Interest coverage ratio (numeric vector)
}
\description{
Calculate Interest Coverage Ratio
}
\keyword{internal}
