% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_margin.R
\name{net_margin}
\alias{net_margin}
\title{Calculate Net Margin}
\usage{
net_margin(net_income, revenue)
}
\arguments{
\item{net_income}{Net income (numeric vector)}

\item{revenue}{Revenue (numeric vector)}
}
\value{
Net margin as decimal (numeric vector)
}
\description{
Calculate Net Margin
}
\keyword{internal}
