% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_margin.R
\name{operating_margin}
\alias{operating_margin}
\title{Calculate Operating Margin}
\usage{
operating_margin(operating_income, revenue)
}
\arguments{
\item{operating_income}{Operating income (numeric vector)}

\item{revenue}{Revenue (numeric vector)}
}
\value{
Operating margin as decimal (numeric vector)
}
\description{
Calculate Operating Margin
}
\keyword{internal}
