% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kpi_data.R
\name{prepare_kpi_data}
\alias{prepare_kpi_data}
\title{Prepare KPI Data for Financial Ratio Charts}
\usage{
prepare_kpi_data(data)
}
\arguments{
\item{data}{Data frame from prepare_fundamentals_data}
}
\value{
Data frame with date and all calculated ratios
}
\description{
Calculates financial ratios from fundamentals data.
}
