% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_display_case.R
\name{to_display_case}
\alias{to_display_case}
\title{Convert snake_case to Display Case}
\usage{
to_display_case(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector in Title Case
}
\description{
Converts strings like "specialty_chemicals" to "Specialty Chemicals".
}
\keyword{internal}
