% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_id.R
\name{get_session_id}
\alias{get_session_id}
\title{Get Shiny Session ID}
\usage{
get_session_id(session)
}
\arguments{
\item{session}{Shiny session object}
}
\value{
Character string with session identifier
}
\description{
Extracts a short identifier from Shiny session for logging.
}
