% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_equities_taxonomy.R
\name{join_equities_taxonomy}
\alias{join_equities_taxonomy}
\title{Join Equities Taxonomy to TTM Data}
\usage{
join_equities_taxonomy(ttm_data)
}
\arguments{
\item{ttm_data}{TTM artifact tibble with \code{industry} column}
}
\value{
TTM data with \code{subsector} column added
}
\description{
Left joins subsector taxonomy onto TTM artifact by industry.
}
